Create README.md
Finding a suitable replacement for simple effect additivity  (SEA) synergy theory in the case of very low flux exposures to ionizing radiation seems to be very hard. If the flux is assumed constant and is used as the main predictor variable adapting incremental effect additivity (IEA) is hopefully feasible. This repository will be used in that attempt